setup: rename folder name (modules -> utils)
